Refactor: [#21]  포스트 디테일 페이지 리팩토링
