+++ components/post/post-detail-body.tsx
@@ -1,0 +1,6 @@
+import { MDXRemote, MDXRemoteProps } from "next-mdx-remote"
+import React from "react"
+
+export default function PostDetailBody(props: MDXRemoteProps) {
+  return <MDXRemote {...props} />
+}
