+++ pages/blog/[category]/[slug]/index.tsx
@@ -1,8 +1,6 @@
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next"
-import Image from "next/image"
 import { getPostDetail, getPostPaths, parsePostAbstract } from "@/lib/post.lib"
 import { serialize } from "next-mdx-remote/serialize"
-import { MDXRemote } from "next-mdx-remote"
 // @ts-expect-error no types
 import remarkA11yEmoji from "@fec/remark-a11y-emoji"
 import remarkGfm from "remark-gfm"
@@ -12,20 +10,15 @@ import rehypeSlug from "rehype-slug"
 import rehypePrettyCode from "rehype-pretty-code"
 import rehypeAutolinkHeadings from "rehype-autolink-headings"
 // import rehypePrism from "rehype-prism-plus"
-import { Badge } from "@/components/ui/badge"
 import PostDetailHeader from "@/components/post/post-detail-header"
 import Giscus from "@/components/giscus"
+import PostDetailBody from "@/components/post/post-detail-body"
 
 export default function PostDetailPage(props: InferGetStaticPropsType<typeof getStaticProps>) {
   return (
     <section className="m-auto w-[90%] prose prose-neutral dark:prose-invert max-w-[680px]">
-      <Badge className="mb-4 px-2.5 py-1">{props.categoryPath.toUpperCase()}</Badge>
       <PostDetailHeader {...props} />
-
-      <div className="relative w-full h-[63.333vw] max-h-[400px] lg:h-[39.063vw] lg:max-h-[450px] not-prose mt-10">
-        <Image src={props.thumbnail} fill alt="thumnnail" objectFit="cover" objectPosition="center" />
-      </div>
-      <MDXRemote {...props.mdxSource} />
+      <PostDetailBody {...props.mdxSource} />
       <Giscus />
     </section>
   )
